simplify CryptStreamCreator
